# swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    BindingAttribute:
        properties:
            Type:
                type: string
            Value:
                type: string
        type: object
        x-go-package: github.com/Knox-AAU/DatabaseLayer_Server/pkg/graph
    GetTriple:
        properties:
            o:
                $ref: '#/definitions/BindingAttribute'
            p:
                $ref: '#/definitions/BindingAttribute'
            s:
                $ref: '#/definitions/BindingAttribute'
        title: GetTriple requires the json tags to match with the queries that are used to retrieve it.
        type: object
        x-go-package: github.com/Knox-AAU/DatabaseLayer_Server/pkg/graph
    PostBody:
        properties:
            triples:
                description: |-
                    Triples is an array of triples.
                    Each triple's first element is the subject, second is the predicate and third is the object.
                    Only accepts exactly 3 elements per triple.
                items:
                    items:
                        type: string
                    type: array
                type: array
                x-go-name: Triples
        required:
            - triples
        type: object
        x-go-package: github.com/Knox-AAU/DatabaseLayer_Server/pkg/graph
    Result:
        properties:
            query:
                type: string
                x-go-name: Query
            triples:
                items:
                    $ref: '#/definitions/GetTriple'
                type: array
                x-go-name: Triples
        type: object
        x-go-package: github.com/Knox-AAU/DatabaseLayer_Server/pkg/http/rest
host: http://192.38.54.90
info:
    description: REST API for the KNOX database.
    title: Database Layer Server API.
paths:
    /triples:
        get:
            description: To query the whole graph, leave parameters empty.
            operationId: getTriples
            parameters:
                - description: Target graph
                  in: query
                  name: g
                  required: true
                  type: string
                - description: Subjects
                  in: query
                  items:
                    type: string
                  name: s
                  type: array
                - description: Objects
                  in: query
                  items:
                    type: string
                  name: o
                  type: array
                - description: Predicates
                  in: query
                  items:
                    type: string
                  name: p
                  type: array
            produces:
                - application/json
            responses:
                "200":
                    description: filtered triples response
                    schema:
                        $ref: '#/definitions/Result'
            summary: This endpoint queries the graph for triples applying filters.
        post:
            description: If a new predicate is sent with an existing subject, will the existing subject be updated with the new predicate.
            operationId: UpsertTriples
            parameters:
                - description: Target graph
                  in: query
                  name: g
                  required: true
                  type: string
                - description: Triples to upsert
                  in: body
                  name: triples
                  required: true
                  schema:
                    $ref: '#/definitions/PostBody'
            produces:
                - application/json
            responses:
                "200":
                    description: response with produced query and null value for triples
                    schema:
                        $ref: '#/definitions/Result'
            summary: This endpoint upserts triples.
produces:
    - application/json
schemes:
    - http
swagger: "2.0"